+++ Vostok.Hosting/Setup/IVostokHerculesServiceDiscoveryEventsSenderBuilder.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using JetBrains.Annotations;
+using Vostok.ServiceDiscovery.Telemetry.Hercules;
+
+namespace Vostok.Hosting.Setup
+{
+    [PublicAPI]
+    public interface IVostokHerculesServiceDiscoveryEventsSenderBuilder
+    {
+        bool IsEnabled { get; }
+
+        IVostokHerculesServiceDiscoveryEventsSenderBuilder Enable();
+        IVostokHerculesServiceDiscoveryEventsSenderBuilder Disable();
+
+        IVostokHerculesServiceDiscoveryEventsSenderBuilder SetStream([NotNull] string stream);
+        IVostokHerculesServiceDiscoveryEventsSenderBuilder SetApiKeyProvider([NotNull] Func<string> apiKeyProvider);
+        IVostokHerculesServiceDiscoveryEventsSenderBuilder CustomizeSettings([NotNull] Action<HerculesServiceDiscoveryEventsSenderSettings> settingsCustomization);
+    }
+}